+++ 2.3-1/src/c/statisticsFunctions/max/u8rowmaxa.c
@@ -9,15 +9,16 @@
  *  http://www.cecill.info/licences/Licence_CeCILL_V2-en.txt
  *
  */
-
 #include "statMax.h"
+//#include "max.h"
 
 void u8rowmaxa(uint8 *in, int rows, int columns, uint8* out) {
   int i = 0, j = 0;
 
   for (i = 0; i < columns; i++) {
-	  out[i]=in[i*rows];
+	  out[i]=(uint8)in[i*rows];
 	  for (j = 1 ; j < rows ; j++)
-     		 if (in[i*rows+j]>out[i]) out[i] = (uint8)in[i*rows+j];
+     		 if (in[i*rows+j]>out[i])
+                   out[i] = (uint8)in[i*rows+j];
     }
 }
